pyproject.toml
[project]
name = "stock-research"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "OpenDartReader>=0.2",
    "FinanceDataReader>=0.9",
    "yfinance>=0.2",
    "pandas-ta>=0.3",
    "edgartools>=2.0",
    "feedparser>=6.0",
    "supabase>=2.0",
    "pydantic>=2.0",
    "tiktoken>=0.7",
    "pytest>=8.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[tool.uv.workspace]
members = ["apps/collector"]

[tool.pytest.ini_options]
testpaths = ["apps/collector/tests"]
python_files = ["test_*.py"]
